全删除注释前

+++ eu-gateway/src/main/java/com/xiaojun/common/http/CookieUtil.java
@@ -58,11 +58,11 @@ public class CookieUtil {
      * @param value    保存值
      * @author jxf
      */
-    public static HttpServletResponse setCookie(HttpServletResponse response, String name, String value, int time) {
+    public static HttpServletResponse setCookie(HttpServletResponse response, String name, String value, long time) {
         return doSetCookie(response, name, value, time, true);
     }
 
-    private static HttpServletResponse doSetCookie(HttpServletResponse response, String name, String value, int time, boolean isHttpOnly) {
+    private static HttpServletResponse doSetCookie(HttpServletResponse response, String name, String value, long time, boolean isHttpOnly) {
         // new一个Cookie对象,键值对为参数
         Cookie cookie = new Cookie(name, value);
         // tomcat下多应用共享
@@ -73,7 +73,7 @@ public class CookieUtil {
         } catch (UnsupportedEncodingException e) {
             log.error("", e);
         }
-        cookie.setMaxAge(time);
+        cookie.setMaxAge((int) time);
         // 将Cookie添加到Response中,使之生效
         response.addCookie(cookie); // addCookie后，如果已经存在相同名字的cookie，则最新的覆盖旧的cookie
         if (isHttpOnly) {
